+++ src/main/java/com/weizujie/attendance/mapper/ClazzMapper.java
@@ -13,7 +13,7 @@ public interface ClazzMapper extends BaseMapper<Clazz> {
 
     int addClazz(Clazz clazz);
 
-    int deleteClazz(List<Long> ids);
+    int deleteClazz(List<Integer> ids);
 
     int editClazz(Clazz clazz);
